prepare for generic observers
