+++ ecs/world_internal.go
@@ -1,6 +1,7 @@
 package ecs
 
 import (
+	"fmt"
 	"reflect"
 )
 
@@ -342,6 +343,9 @@ func (w *World) emitEvent(e *Event, entity Entity) {
 		}
 		table := w.storage.entities[entity.id].table
 		mask = &w.storage.archetypes[w.storage.tables[table].archetype].mask
+	}
+	if e.hasComp && !mask.Get(e.component.id) {
+		panic(fmt.Sprintf("event entity does not have component with ID %d", e.component.id))
 	}
 	w.storage.observers.FireCustom(e.eventType, entity, e.component, e.hasComp, mask)
 }
